@@ -7,17 +7,17 @@
   "repository": "@elrondnetwork/dapp-template",
   "dependencies": {
     "@elrondnetwork/dapp-core": "1.2.7",
-    "@elrondnetwork/dapp-utils": "0.1.3",
-    "@elrondnetwork/erdjs": "9.0.3",
+    "@elrondnetwork/erdjs": "10.2.5",
+    "@elrondnetwork/erdjs-network-providers": "0.1.5",
     "@fortawesome/fontawesome-svg-core": "6.1.0",
     "@fortawesome/free-solid-svg-icons": "6.1.0",
     "@fortawesome/react-fontawesome": "0.1.18",
     "axios": "0.21.1",
     "bootstrap": "4.6.0",
     "moment": "2.29.1",
-    "react": "17.0.2",
+    "react": "18.1.0",
     "react-bootstrap": "2.1.2",
-    "react-dom": "17.0.2",
+    "react-dom": "18.1.0",
     "react-router-dom": "6.2.1",
     "react-scripts": "4.0.3",
     "sass": "1.52.1",
@@ -25,7 +25,7 @@
     "web-vitals": "1.0.1"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "HTTPS=true PORT=3001 react-scripts start",
     "start-testnet": "npm run copy-testnet-config & npm run start",
     "start-devnet": "npm run copy-devnet-config & npm run start",
     "build": "react-scripts build",
@@ -61,8 +61,8 @@
     "@testing-library/user-event": "13.5.0",
     "@types/jest": "26.0.15",
     "@types/node": "12.0.0",
-    "@types/react": "17.0.7",
-    "@types/react-dom": "17.0.0",
+    "@types/react": "18.0.10",
+    "@types/react-dom": "18.0.5",
     "@types/react-router-dom": "5.1.7",
     "@typescript-eslint/eslint-plugin": "4.22.1",
     "@typescript-eslint/parser": "4.22.1",
@@ -78,6 +78,7 @@
     "eslint-plugin-promise": "4.2.1",
     "eslint-plugin-react": "7.22.0",
     "eslint-plugin-react-hooks": "4.2.0",
-    "prettier": "2.3.2"
+    "prettier": "2.3.2",
+    "react-error-overlay": "^6.0.9"
   }
 }
